hiretech/companies/search_indexs.py
from haystack import indexes

from .models import Company

class CompanyIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)

    def get_model(self):
        return Company
