requirements.txt
psycopg2
elasticsearch
python-coveralls
Django==1.8.4
Pillow==2.9.0
wheel==0.24.0
django-haystack==2.4.0
django-libsass
django-markdown-deux
django-nocaptcha-recaptcha
django-taggit
django-suit==0.2.13
django-imagekit
https://github.com/kronok/django-drip/archive/cd7dcabc129a5b2857548e4e181addf0b4eaba8a.zip
